requirements.txt
Django>=4.2.0,<5.0
Pillow>=10.0.0
python-decouple>=3.8
django-ckeditor>=6.7.0
django-cleanup>=8.0.0